+++ src/index.ts
@@ -2,10 +2,11 @@
 import type {GeoJSONVTFeature, GeoJSONVTFeatureNonPoint, GeoJSONVTFeaturePoint, GeoJSONVTTile} from './transform';
 import type {GeoJSONVTInternalTile, GeoJSONVTInternalTileFeature, GeoJSONVTInternalTileFeaturePoint, GeoJSONVTInternalTileFeatureNonPoint} from './tile';
 import type {GeoJSONVTFeatureDiff, GeoJSONVTSourceDiff} from './difference';
-import type {GeoJSONVTInternalFeature, GeoJSONVTInternalLineStringFeature, GeoJSONVTInternalMultiLineStringFeature, GeoJSONVTInternalMultiPointFeature, GeoJSONVTInternalMultiPolygonFeature, GeoJSONVTInternalPointFeature, GeoJSONVTInternalPolygonFeature, GeoJSONVTOptions, PartialGeoJSONVTFeature, StartEndSizeArray} from './definitions';
+import type {GeoJSONVTInternalFeature, GeoJSONVTInternalLineStringFeature, GeoJSONVTInternalMultiLineStringFeature, GeoJSONVTInternalMultiPointFeature, GeoJSONVTInternalMultiPolygonFeature, GeoJSONVTInternalPointFeature, GeoJSONVTInternalPolygonFeature, GeoJSONVTOptions, GeoJSONToTileOptions, PartialGeoJSONVTFeature, StartEndSizeArray} from './definitions';
 import type {SuperclusterTile, SuperclusterTileFeature, SuperclusterOptions, ClusterFeature, ClusterProperties, KDBushWithData} from './supercluster';
 import {GeoJSONVT} from './geojsonvt';
 import {Supercluster} from './supercluster';
+import {geoJSONToTile} from './geoJSONToTile';
 
 function geojsonvt(data: GeoJSON.GeoJSON, options?: GeoJSONVTOptions) {
     return new GeoJSONVT(data, options);
@@ -13,13 +14,15 @@ function geojsonvt(data: GeoJSON.GeoJSON, options?: GeoJSONVTOptions) {
 
 geojsonvt.GeoJSONVT = GeoJSONVT;
 geojsonvt.Supercluster = Supercluster;
+geojsonvt.geoJSONToTile = geoJSONToTile;
 
 export default geojsonvt;
 
 export type {
     GeoJSONVT,
     GeoJSONVTInternalFeature, 
-    GeoJSONVTOptions, 
+    GeoJSONVTOptions,
+    GeoJSONToTileOptions,
     GeoJSONVTInternalTile, 
     GeoJSONVTInternalTileFeature, 
     PartialGeoJSONVTFeature,